add exceptions, change log level

+++ src/main/java/dao/implementation/mySQL/SubjectDaoImpl.java
@@ -1,7 +1,12 @@
-package dao;
+package dao.implementation.mySQL;
 
-import dao.impl.mysql.ISubjectDao;
+import dao.ISubjectDao;
+import entity.Group;
+import entity.Student;
 import entity.Subject;
+import exceptions.EntityExistsException;
+import exceptions.EntityNotFoundException;
+import exceptions.InvalidIdException;
 import org.apache.log4j.Logger;
 import utils.HibernateUtils;
 import utils.Utils;
@@ -25,90 +30,103 @@ public class SubjectDaoImpl implements ISubjectDao {
 		this.entityManager = entityManager;
 	}
 
-	public Subject create(Subject subject) {
-		LOG.info("User try to create subject");
+	public Subject create(Subject subject) throws EntityExistsException {
+		LOG.trace("User try to create subject");
+
+		if (entityManager.find(Subject.class, subject.getId()) != null) {
+			throw new EntityExistsException(subject.toString());
+		}
+
 		EntityTransaction transaction = entityManager.getTransaction();
 		try {
 			transaction.begin();
-			LOG.info("Transaction began");
+			LOG.trace("Transaction began");
 			entityManager.persist(subject);
-			LOG.info("Subject persisted");
+			LOG.trace("Subject persisted");
 			transaction.commit();
-			LOG.info("Transaction commited");
+			LOG.trace("Transaction commited");
 		} catch (Exception e) {
 			LOG.error("Operation failed");
 			transaction.rollback();
-			LOG.info("Transaction rolled back");
+			LOG.trace("Transaction rolled back");
 			return null;
 		}
 		return subject;
 	}
 
-	public boolean delete(Subject subject) {
-		LOG.info("User try to delete group");
+	public boolean delete(Subject subject) throws EntityNotFoundException {
+		LOG.trace("User try to delete group");
+
+		if (entityManager.find(Subject.class, subject.getId()) == null) {
+			throw new EntityNotFoundException(subject.toString());
+		}
+
 		EntityTransaction transaction = entityManager.getTransaction();
 		try {
 			transaction.begin();
-			LOG.info("Transaction began");
+			LOG.trace("Transaction began");
 			subject = entityManager.find(Subject.class, subject.getId());
-			LOG.info("Add subject into managed context");
+			LOG.trace("Add subject into managed context");
 			entityManager.remove(subject);
-			LOG.info("Group removed");
+			LOG.trace("Group removed");
 			transaction.commit();
-			LOG.info("Transaction commited");
+			LOG.trace("Transaction commited");
 		} catch (Exception e) {
 			LOG.error("Operation failed");
 			transaction.rollback();
-			LOG.info("Transaction rolled back");
+			LOG.trace("Transaction rolled back");
 			return false;
 		}
 		return true;
 	}
 
-	public boolean update(Subject subject) {
-		LOG.info("User try to update subject");
+	public boolean update(Subject subject) throws EntityNotFoundException {
+		LOG.trace("User try to update subject");
 		EntityTransaction transaction = entityManager.getTransaction();
 
 		Subject subjectFromDB = entityManager.find(Subject.class, subject.getId());
-		LOG.info("Subject found correct");
+
+		if (subjectFromDB == null) {
+			throw new EntityNotFoundException(subject.toString());
+		}
+
+		LOG.trace("Subject found correct");
 		Utils.copySubject(subject, subjectFromDB);
-		LOG.info("Subject data updated correct");
+		LOG.trace("Subject data updated correct");
 		try {
 			transaction.begin();
-			LOG.info("Transaction began");
+			LOG.trace("Transaction began");
 			entityManager.persist(subjectFromDB);
-			LOG.info("Subject added to managed context");
+			LOG.trace("Subject added to managed context");
 			transaction.commit();
-			LOG.info("Transaction commited");
+			LOG.trace("Transaction commited");
 		} catch (Exception e) {
 			LOG.error("Operation failed");
 			transaction.rollback();
-			LOG.info("Transaction rolled back");
+			LOG.trace("Transaction rolled back");
 			return false;
 		}
 		return true;
 	}
 
-	public Subject findById(Object id) {
-		LOG.info("User try to find subject by id");
-		try {
-			Subject subjectFromDB = entityManager.find(Subject.class, id);
-			if (subjectFromDB != null)
-			{
-				LOG.info("Group " + subjectFromDB + " found correct");
-			} else {
-				LOG.info("Group with id " + id + " not found");
-			}
-			return subjectFromDB;
-		} catch (Exception e) {
-			LOG.error("Operation failed");
-			return null;
+	public Subject findById(Object id) throws InvalidIdException, EntityNotFoundException {
+		LOG.trace("User try to find subject by id");
+
+		if (!(id instanceof Integer) || ((Integer)id < 0)) {
+			throw new InvalidIdException(id.toString());
 		}
+
+		Subject subjectFromDB = entityManager.find(Subject.class, id);
+		if (subjectFromDB == null) {
+			throw new EntityNotFoundException((Integer)id);
+		}
+
+		return subjectFromDB;
 	}
 
 	@Override
 	public List<Subject> getAll(int firstRow, int rowsAmount) {
-		LOG.info("User try to find subjects from " + firstRow + " row and with row amount " + rowsAmount);
+		LOG.trace("User try to find subjects from " + firstRow + " row and with row amount " + rowsAmount);
 		return entityManager.createQuery(GET_ALL_SUBJECTS)
 				.setFirstResult(firstRow)
 				.setMaxResults(rowsAmount)
@@ -116,15 +134,16 @@ public class SubjectDaoImpl implements ISubjectDao {
 	}
 
 	@Override
-	public List<Subject> getGumanitariumSubjects() {
-		LOG.info("User try to find only gumanitarium subjects");
+	public List<Subject> getGumanitariumSubjects() throws EntityNotFoundException {
+		LOG.trace("User try to find only gumanitarium subjects");
 		List<Subject> gumanitariumSubjects = entityManager.createQuery(GET_GUMANITARIUM_SUBJECTS).getResultList();
 		if (gumanitariumSubjects != null) {
-			LOG.info("Gumanitarium subjects are found correctly");
+			LOG.trace("Gumanitarium subjects are found correctly");
 			return gumanitariumSubjects;
+		} else {
+			throw new EntityNotFoundException("Gumanitarium subjects ");
 		}
-		LOG.info("Gumanitarium subjects are not found");
-		return null;
+
 
 	}
